fixed ValueProvider tests
